+++ packages/client/hooks/useTipTapTypingStatus.ts
@@ -3,13 +3,17 @@ import {useEffect, useRef, useState} from 'react'
 import EditCommentingMutation from '../mutations/EditCommentingMutation'
 import useAtmosphere from './useAtmosphere'
 
-export const useTipTapTypingStatus = (editor: Editor | null, discussionId: string) => {
+export const useTipTapTypingStatus = (
+  editor: Editor | null,
+  discussionId: string,
+  isAnonymous: boolean
+) => {
   const [isTyping, setIsTyping] = useState<boolean | null>(null)
   const timeoutRef = useRef<NodeJS.Timeout | null>(null)
   const atmosphere = useAtmosphere()
 
   useEffect(() => {
-    if (!editor) return
+    if (!editor || isAnonymous) return
     const handleUpdate = () => {
       if (editor.isEmpty) return
       setIsTyping(true)
@@ -32,7 +36,7 @@ export const useTipTapTypingStatus = (editor: Editor | null, discussionId: strin
       editor.off('blur', handleBlur)
       if (timeoutRef.current) clearTimeout(timeoutRef.current)
     }
-  }, [editor])
+  }, [editor, isAnonymous])
 
   useEffect(() => {
     if (isTyping === null) return